+++ plugins/clickhouse/src/components/panel/LogsDocument.tsx
@@ -1,52 +1,103 @@
+import React, { useState } from 'react';
 import { TableText, Td, Tr } from '@patternfly/react-table';
-import { DropdownPosition } from '@patternfly/react-core';
-import React from 'react';
 
-import Details from './details/Details';
 import { IDocument } from '../../utils/interfaces';
+import LogsDocumentDetails from './LogsDocumentDetails';
 import { formatTimeWrapper } from '../../utils/helpers';
 
 interface ILogsDocumentProps {
   document: IDocument;
   fields?: string[];
-  showDetails?: (details: React.ReactNode) => void;
 }
 
-const LogsDocument: React.FunctionComponent<ILogsDocumentProps> = ({
-  document,
-  fields,
-  showDetails,
-}: ILogsDocumentProps) => {
+const LogsDocument: React.FunctionComponent<ILogsDocumentProps> = ({ document, fields }: ILogsDocumentProps) => {
+  const [isExpanded, setIsExpanded] = useState<boolean>(false);
+
   const defaultActions = [
     {
-      onClick: (): void =>
-        showDetails
-          ? showDetails(<Details document={document} close={(): void => showDetails(undefined)} />)
-          : undefined,
-      title: 'Details',
+      title: (
+        <a
+          style={{ color: 'inherit', textDecoration: 'inherit' }}
+          href={URL.createObjectURL(new Blob([JSON.stringify({ data: document }, null, 2)]))}
+          download={`${document['timestamp']}__${document['container_name']}__${document['pod_name']}__${document['namespace']}__${document['cluster']}.json`}
+        >
+          Download JSON
+        </a>
+      ),
     },
   ];
 
   return (
-    <Tr>
-      <Td
-        noPadding={true}
-        style={{ padding: 0 }}
-        actions={{ dropdownPosition: DropdownPosition.left, items: defaultActions }}
-      />
-      <Td dataLabel="Time">
-        <TableText wrapModifier="nowrap">{formatTimeWrapper(document['timestamp'])}</TableText>
-      </Td>
-      {fields && fields.length > 0 ? (
-        fields.map((field, index) => (
-          <Td key={index} dataLabel={field}>
-            {document[field]}
+    <React.Fragment>
+      <Tr>
+        <Td
+          noPadding={true}
+          style={{ padding: 0 }}
+          expand={{ isExpanded: isExpanded, onToggle: (): void => setIsExpanded(!isExpanded), rowIndex: 0 }}
+        />
+        <Td className="pf-u-text-wrap pf-u-text-break-word" dataLabel="Time">
+          <TableText wrapModifier="nowrap">{formatTimeWrapper(document['timestamp'])}</TableText>
+        </Td>
+        {fields && fields.length > 0 ? (
+          fields.map((field, index) => (
+            <Td key={index} className="pf-u-text-wrap pf-u-text-break-word" dataLabel={field}>
+              {document[field]}
+            </Td>
+          ))
+        ) : (
+          <Td className="pf-u-text-wrap pf-u-text-break-word" dataLabel="Log">
+            <div className="kobsio-clickhouse-logs-preview">
+              <span className="pf-u-mr-sm pf-u-mb-sm">
+                <span className="pf-u-background-color-200 pf-u-p-xs">cluster:</span>
+                <span className="pf-u-p-xs"> {document['cluster']}</span>
+              </span>
+              <span className="pf-u-mr-sm pf-u-mb-sm">
+                <span className="pf-u-background-color-200 pf-u-p-xs">namespace:</span>
+                <span className="pf-u-p-xs"> {document['namespace']}</span>
+              </span>
+              <span className="pf-u-mr-sm pf-u-mb-sm">
+                <span className="pf-u-background-color-200 pf-u-p-xs">app:</span>
+                <span className="pf-u-p-xs"> {document['app']}</span>
+              </span>
+              <span className="pf-u-mr-sm pf-u-mb-sm">
+                <span className="pf-u-background-color-200 pf-u-p-xs">pod_name:</span>
+                <span className="pf-u-p-xs"> {document['pod_name']}</span>
+              </span>
+              <span className="pf-u-mr-sm pf-u-mb-sm">
+                <span className="pf-u-background-color-200 pf-u-p-xs">container_name:</span>
+                <span className="pf-u-p-xs"> {document['container_name']}</span>
+              </span>
+              <span className="pf-u-mr-sm pf-u-mb-sm">
+                <span className="pf-u-background-color-200 pf-u-p-xs">host:</span>
+                <span className="pf-u-p-xs"> {document['host']}</span>
+              </span>
+
+              {Object.keys(document)
+                .filter((key) => key.startsWith('content.') && document[key].length < 128)
+                .map((key) => (
+                  <span key={key} className="pf-u-mr-sm pf-u-mb-sm">
+                    <span className="pf-u-background-color-200 pf-u-p-xs">{key}:</span>
+                    <span className="pf-u-p-xs"> {document[key]}</span>
+                  </span>
+                ))}
+
+              <span className="pf-u-mr-sm pf-u-mb-sm">
+                <span className="pf-u-background-color-200 pf-u-p-xs">log:</span>
+                <span className="pf-u-p-xs"> {document['log']}</span>
+              </span>
+            </div>
           </Td>
-        ))
-      ) : (
-        <Td dataLabel="Log">{document['log']}</Td>
-      )}
-    </Tr>
+        )}
+        <Td noPadding={true} style={{ padding: 0 }} actions={{ items: defaultActions }} />
+      </Tr>
+      <Tr isExpanded={isExpanded}>
+        <Td />
+        <Td colSpan={fields && fields.length > 0 ? fields.length : 2}>
+          {isExpanded && <LogsDocumentDetails document={document} />}
+        </Td>
+        <Td />
+      </Tr>
+    </React.Fragment>
   );
 };